+++ src/titiler/core/titiler/core/dependencies.py
@@ -9,10 +9,9 @@ from typing import Dict, List, Optional, Sequence, Tuple, Union
 import numpy
 from fastapi import HTTPException, Query
 from rasterio.crs import CRS
-from rasterio.enums import Resampling
 from rio_tiler.colormap import cmap, parse_color
 from rio_tiler.errors import MissingAssets, MissingBands
-from rio_tiler.types import ColorMapType
+from rio_tiler.types import ColorMapType, RIOResampling
 
 if sys.version_info >= (3, 9):
     from typing import Annotated  # pylint: disable=no-name-in-module
@@ -22,9 +21,6 @@ else:
 
 ColorMapName = Enum(  # type: ignore
     "ColorMapName", [(a, a) for a in sorted(cmap.list())]
-)
-ResamplingName = Enum(  # type: ignore
-    "ResamplingName", [(r.name, r.name) for r in Resampling]
 )
 
 
@@ -349,18 +345,18 @@ class DatasetParams(DefaultDependency):
         ),
     ] = False
     resampling_method: Annotated[
-        ResamplingName,
+        RIOResampling,
         Query(
             alias="resampling",
             description="Resampling method.",
         ),
-    ] = ResamplingName.nearest  # type: ignore
+    ] = "nearest"
 
     def __post_init__(self):
         """Post Init."""
         if self.nodata is not None:
             self.nodata = numpy.nan if self.nodata == "nan" else float(self.nodata)
-        self.resampling_method = self.resampling_method.value  # type: ignore
+        self.resampling_method = self.resampling_method
 
 
 @dataclass
@@ -385,7 +381,7 @@ def RescalingParams(
         Query(
             title="Min/Max data Rescaling",
             description="comma (',') delimited Min,Max range. Can set multiple time for multiple bands.",
-            example=["0,2000", "0,1000", "0,10000"],  # band 1  # band 2  # band 3
+            examples=["0,2000", "0,1000", "0,10000"],  # band 1  # band 2  # band 3
         ),
     ] = None,
 ) -> Optional[RescaleType]:
@@ -410,7 +406,7 @@ class StatisticsParams(DefaultDependency):
             alias="c",
             title="Pixels values for categories.",
             description="List of values for which to report counts.",
-            example=[1, 2, 3],
+            examples=[1, 2, 3],
         ),
     ] = None
     percentiles: Annotated[
@@ -418,8 +414,8 @@ class StatisticsParams(DefaultDependency):
         Query(
             alias="p",
             title="Percentile values",
-            description="List of percentile values.",
-            example=[2, 5, 95, 98],
+            description="List of percentile values (default to [2, 98]).",
+            examples=[2, 5, 95, 98],
         ),
     ] = None
 
@@ -473,7 +469,7 @@ range affects the automatic bin computation as well.
 
 link: https://numpy.org/doc/stable/reference/generated/numpy.histogram.html
             """,
-            example="0,1000",
+            examples="0,1000",
         ),
     ] = None