+++ src/titiler/mosaic/titiler/mosaic/resources/__init__.py
@@ -1,1 +1,0 @@
-"""titiler.core.resources."""